# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIOps Platform API
  version: 1.0.0
  description: >-
    Versioned HTTP contract for the AIOps platform. Sensitive credentials,
    passwords, tokens and Kubernetes evidence payloads are intentionally not
    represented in response schemas.
servers:
  - url: /
...
paths:
...
  /api/v1/fleet/resources/search/filters:
    get:
      tags: [fleet]
      summary: List the current user's saved global-search filters
      operationId: listGlobalSearchFilters
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User-owned saved filters and the fixed per-user limit, content: { application/json: { schema: { $ref: '#/components/schemas/SavedGlobalSearchFilterList' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    post:
      tags: [fleet]
      summary: Save the current bounded global-search query for the current user
      operationId: createGlobalSearchFilter
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SavedGlobalSearchFilterCreate' } } } }
      responses:
        '201': { description: Saved filter, content: { application/json: { schema: { $ref: '#/components/schemas/SavedGlobalSearchFilter' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/fleet/resources/search/filters/{filter_id}:
    patch:
      tags: [fleet]
      summary: Rename a saved filter or overwrite it with a complete bounded query
      operationId: updateGlobalSearchFilter
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/FilterID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SavedGlobalSearchFilterUpdate' } } } }
      responses:
        '200': { description: Updated saved filter, content: { application/json: { schema: { $ref: '#/components/schemas/SavedGlobalSearchFilter' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
    delete:
      tags: [fleet]
      summary: Delete one of the current user's saved global-search filters
      operationId: deleteGlobalSearchFilter
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/FilterID' }]
      responses:
        '204': { description: Saved filter deleted }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
  /api/v1/auth/login:
    post: { tags: [auth], summary: Create access and refresh session, operationId: login, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }, responses: { '200': { $ref: '#/components/responses/Login' }, '401': { $ref: '#/components/responses/Error' } } }
  /api/v1/auth/refresh:
    post: { tags: [auth], summary: Rotate refresh cookie and issue access token, operationId: refresh, security: [{ refreshCookie: [] }], responses: { '200': { $ref: '#/components/responses/Login' }, '401': { $ref: '#/components/responses/Error' } } }
  /api/v1/auth/logout:
    post: { tags: [auth], summary: Revoke current refresh session, operationId: logout, security: [{ refreshCookie: [] }], responses: { '204': { description: Revoked } } }
...
  /api/v1/auth/password-change:
    post: { tags: [auth], summary: Change own password, operationId: changePassword, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordChangeRequest' } } } }, responses: { '204': { description: Password changed and sessions revoked }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/auth/sessions:
    get: { tags: [auth], summary: List own refresh sessions, operationId: sessions, security: [{ bearerAuth: [], refreshCookie: [] }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /api/v1/auth/sessions/{session_id}:
    delete: { tags: [auth], summary: Revoke another own session, operationId: revokeSession, security: [{ bearerAuth: [], refreshCookie: [] }], parameters: [{ $ref: '#/components/parameters/SessionID' }], responses: { '204': { description: Revoked }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/auth/sessions/revoke-others:
    post: { tags: [auth], summary: Revoke all other sessions, operationId: revokeOtherSessions, security: [{ bearerAuth: [], refreshCookie: [] }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /api/v1/users:
    get: { tags: [users], summary: List users, operationId: listUsers, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    post: { tags: [users], summary: Create user, operationId: createUser, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserWrite' } } } }, responses: { '201': { $ref: '#/components/responses/User' }, '409': { $ref: '#/components/responses/Error' } } }
...
  /api/v1/users/{user_id}:
    patch: { tags: [users], summary: Update user status, display name or roles, operationId: updateUser, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/UserID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserPatch' } } } }, responses: { '200': { $ref: '#/components/responses/User' }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/users/{user_id}/password-reset:
    post: { tags: [users], summary: Reset another user's password, operationId: resetPassword, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/UserID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordResetRequest' } } } }, responses: { '204': { description: Reset and sessions revoked }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/clusters:
    get: { tags: [clusters], summary: List registered clusters, operationId: listClusters, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    post: { tags: [clusters], summary: Register encrypted kubeconfig, operationId: createCluster, security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClusterWrite' } } } }, responses: { '201': { $ref: '#/components/responses/Ok' }, '400': { $ref: '#/components/responses/Error' } } }
  /api/v1/clusters/{cluster_id}:
    get: { tags: [clusters], summary: Cluster detail, operationId: getCluster, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    patch: { tags: [clusters], summary: Enable or disable cluster, operationId: updateCluster, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], responses: { '204': { description: Updated } } }
    delete: { tags: [clusters], summary: Delete cluster and credential, operationId: deleteCluster, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], responses: { '204': { description: Deleted } } }
  /api/v1/clusters/{cluster_id}/credentials:
    put: { tags: [clusters], summary: Atomically replace encrypted kubeconfig, operationId: rotateClusterCredentials, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CredentialRotationRequest' } } } }, responses: { '204': { description: Replaced and probe state reset }, '400': { $ref: '#/components/responses/Error' } } }
  /api/v1/clusters/{cluster_id}/probe:
    post: { tags: [clusters], summary: Probe cluster connectivity, operationId: probeCluster, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
...
  /api/v1/clusters/{cluster_id}/diagnoses:
    post: { tags: [diagnoses], summary: Run deterministic diagnosis, operationId: createDiagnosis, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DiagnosisRequest' } } } }, responses: { '201': { $ref: '#/components/responses/Ok' }, '422': { $ref: '#/components/responses/Error' } } }
...
  /api/v1/diagnoses/{diagnosis_id}:
    get: { tags: [diagnoses], summary: Diagnosis detail, operationId: getDiagnosis, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    patch: { tags: [diagnoses], summary: Transition diagnosis status, operationId: transitionDiagnosis, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], responses: { '204': { description: Updated }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/diagnoses/{diagnosis_id}/assignment:
    patch: { tags: [diagnoses], summary: Assign diagnosis, operationId: assignDiagnosis, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], responses: { '204': { description: Assigned } } }
  /api/v1/diagnoses/{diagnosis_id}/feedback:
    post: { tags: [diagnoses], summary: Submit rule feedback, operationId: diagnosisFeedback, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], responses: { '204': { description: Recorded } } }
...
  /api/v1/diagnoses/{diagnosis_id}/remediations/preview:
    post: { tags: [remediation], summary: Preview a confirmed deployment restart, operationId: previewRemediation, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RemediationPreviewRequest' } } } }, responses: { '201': { $ref: '#/components/responses/Ok' }, '409': { $ref: '#/components/responses/Error' } } }
...
  /api/v1/clusters/{cluster_id}/operations/preview:
    post: { tags: [remediation], summary: Dry-run a fixed resource operation, operationId: previewControlledOperation, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ClusterID' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ControlledOperationPreviewRequest' } } } }, responses: { '201': { $ref: '#/components/responses/Ok' }, '400': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } } }
  /api/v1/remediations/{remediation_id}/execute:
    post: { tags: [remediation], summary: Execute a confirmed remediation plan, operationId: executeRemediation, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/RemediationID' }, { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 128 } }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RemediationExecuteRequest' } } } }, responses: { '200': { $ref: '#/components/responses/Ok' }, '409': { $ref: '#/components/responses/Error' }, '502': { $ref: '#/components/responses/Error' } } }
  /api/v1/diagnoses/{diagnosis_id}/explanations:
    get: { tags: [ai], summary: List cited AI explanations, operationId: listExplanations, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    post: { tags: [ai], summary: Generate cited AI explanation, operationId: generateExplanation, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/DiagnosisID' }], responses: { '201': { $ref: '#/components/responses/Ok' }, '429': { $ref: '#/components/responses/Error' } } }
...
  /api/v1/ai/explanations/{explanation_id}/feedback:
    post: { tags: [ai], summary: Rate one AI explanation, operationId: aiFeedback, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/ExplanationID' }], responses: { '204': { description: Recorded }, '409': { $ref: '#/components/responses/Error' } } }
...
  /api/v1/audit-logs/export:
    get: { tags: [audit], summary: Export safe UTF-8 CSV, operationId: exportAuditLogs, security: [{ bearerAuth: [] }], responses: { '200': { description: CSV export, content: { text/csv: { schema: { type: string } } } } } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: aiops_refresh }
  parameters:
    ClusterID: { name: cluster_id, in: path, required: true, schema: { type: integer, format: int64 } }
    UserID: { name: user_id, in: path, required: true, schema: { type: integer, format: int64 } }
    SessionID: { name: session_id, in: path, required: true, schema: { type: integer, format: int64 } }
    DiagnosisID: { name: diagnosis_id, in: path, required: true, schema: { type: integer, format: int64 } }
    ExplanationID: { name: explanation_id, in: path, required: true, schema: { type: integer, format: int64 } }
...
    RemediationID: { name: remediation_id, in: path, required: true, schema: { type: string, format: uuid } }
    FilterID: { name: filter_id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
...
  responses:
    Ok: { description: Successful JSON response, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
    User: { description: User response without password material, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
    Login: { description: Login or refresh response, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
    Error: { description: Structured error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    SavedGlobalSearchFilterCreate:
      type: object
      additionalProperties: false
      required: [name, query]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        query: { type: string, minLength: 2, maxLength: 64 }
        namespace: { type: string, maxLength: 63, default: '' }
        kinds: { type: array, minItems: 1, maxItems: 4, uniqueItems: true, description: Omission selects all four supported kinds, items: { type: string, enum: [Pod, Deployment, Service, Ingress] } }
    SavedGlobalSearchFilterUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      description: Name may be changed alone. Query, namespace and kinds must be supplied together when overwriting the query.
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        query: { type: string, minLength: 2, maxLength: 64 }
        namespace: { type: string, maxLength: 63 }
        kinds: { type: array, minItems: 1, maxItems: 4, uniqueItems: true, items: { type: string, enum: [Pod, Deployment, Service, Ingress] } }
    SavedGlobalSearchFilter:
      type: object
      required: [id, name, query, kinds, schema_version, compatible, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        query: { type: string }
        namespace: { type: string }
        kinds: { type: array, items: { type: string, enum: [Pod, Deployment, Service, Ingress] } }
        schema_version: { type: integer }
        compatible: { type: boolean }
        incompatibility_code: { type: string, enum: [SCHEMA_VERSION, QUERY_SHAPE] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SavedGlobalSearchFilterList:
      type: object
      required: [items, total, limit]
      properties:
        items: { type: array, maxItems: 20, items: { $ref: '#/components/schemas/SavedGlobalSearchFilter' } }
        total: { type: integer, minimum: 0, maximum: 20 }
        limit: { type: integer, enum: [20] }
...
    Error: { type: object, required: [code, message, request_id], properties: { code: { type: string }, message: { type: string }, request_id: { type: string } } }
    LoginRequest: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string, format: password } } }
    LoginResponse: { type: object, required: [access_token, token_type, expires_in, user], properties: { access_token: { type: string }, token_type: { type: string, example: Bearer }, expires_in: { type: integer }, user: { $ref: '#/components/schemas/User' } } }
    User: { type: object, properties: { id: { type: integer }, username: { type: string }, display_name: { type: string }, roles: { type: array, items: { type: string } }, status: { type: string } } }
    UserWrite: { type: object, required: [username, display_name, password, roles], properties: { username: { type: string }, display_name: { type: string }, password: { type: string, format: password }, roles: { type: array, items: { type: string } } } }
    UserPatch: { type: object, properties: { display_name: { type: string }, status: { type: string, enum: [active, disabled] }, roles: { type: array, items: { type: string } } } }
    PasswordChangeRequest: { type: object, required: [current_password, new_password], properties: { current_password: { type: string, format: password }, new_password: { type: string, format: password } } }
    PasswordResetRequest: { type: object, required: [password], properties: { password: { type: string, format: password } } }
    ClusterWrite: { type: object, required: [name, kubeconfig], properties: { name: { type: string }, kubeconfig: { type: string, writeOnly: true } } }
    CredentialRotationRequest: { type: object, required: [kubeconfig], properties: { kubeconfig: { type: string, writeOnly: true } } }
    DiagnosisRequest: { type: object, required: [resource_kind, name], properties: { resource_kind: { type: string, enum: [Pod, Service, Node, Deployment, Ingress, PersistentVolumeClaim, HorizontalPodAutoscaler] }, namespace: { type: string, description: Required for namespaced resources; empty for Node }, name: { type: string } } }
    RemediationPreviewRequest: { type: object, required: [action, target_name], properties: { action: { type: string, enum: [deployment.rollout_restart] }, target_name: { type: string, maxLength: 253 } } }
    ControlledOperationPreviewRequest:
      oneOf:
        - $ref: '#/components/schemas/DeploymentScaleOperation'
        - $ref: '#/components/schemas/CronJobScheduleOperation'
    DeploymentScaleOperation: { type: object, additionalProperties: false, required: [action, namespace, target_name, desired_replicas], properties: { action: { type: string, enum: [deployment.scale] }, namespace: { type: string, maxLength: 63 }, target_name: { type: string, maxLength: 253 }, desired_replicas: { type: integer, minimum: 0, maximum: 1000 } } }
    CronJobScheduleOperation: { type: object, additionalProperties: false, required: [action, namespace, target_name], properties: { action: { type: string, enum: [cronjob.suspend, cronjob.resume] }, namespace: { type: string, maxLength: 63 }, target_name: { type: string, maxLength: 253 } } }
    RemediationExecuteRequest: { type: object, required: [confirmation_token], properties: { confirmation_token: { type: string, writeOnly: true } } }